requirements-rf31.txt
docutils
ipykernel
ipywidgets
lunr
nbformat==4.4.0
nbimporter
pillow
pygments
robotframework==3.1.2

nbsphinx
sphinx
sphinx-rtd-theme

pytest
pytest-runner

black
coverage==4.5.4
coveralls
isort
pdbpp
pydocstyle
pylama
pylint

faker
iplantuml
jsonschema
jupyter
jupyter-starters==0.2.2a0
jupyter-contrib-nbextensions
jupyterlab
jupytext
requests
RESTinstance
rise
robotframework-appiumlibrary
robotframework-faker
robotframework-seleniumlibrary
robotframework-seleniumtestability
robotframework-seleniumscreenshots
robotframework-jupyterlibrary

text-unidecode==1.3
prompt-toolkit==2.0.10

Appium-Python-Client
